# edulingo/swagger.yaml
openapi: 3.0.0
info:
  title: Edulingo API
  version: 1.0.0
  description: |
    Edulingo backend API documentation (users, auth, courses, quizzes, coupons, certificates, feedbacks).
  contact:
    name: Project Team
    email: team@edulingo.local
servers:
  - url: http://localhost:9999
    description: Local development server (default port from app.js)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        role:
          type: string
        token:
          type: string
...
    UserRegister:
      type: object
      required:
        - full_name
        - email
        - password
      properties:
        full_name:
          type: string
          example: "Nguyen Van A"
        email:
          type: string
          format: email
          example: "a@example.com"
        password:
          type: string
          format: password
          example: "secret123"
        name:
          type: string
          example: "nguyena"
        gender:
          type: string
          enum: [male, female, other]
          example: male
    UserResponse:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        email:
          type: string
        role:
          type: string
paths:
  /api/auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/logout:
    post:
      summary: Logout (client should remove token)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
  /api/auth/register:
    post:
      summary: Register new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Đăng ký thành công"
                  role:
                    type: string
                    example: "learner"
                  user:
                    $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation error or missing fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  error:
                    type: string
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email already in use"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  error:
                    type: string
...